Arceus' true nature
New Ability: Divine Grace:
On switch-in, Arceus type changes to the most effective type for the situation. Requires Legend Plate to actually activate.
Psychosis: The user gains a 1.5x increase to its attacking stats but loses the ability to use Status moves.
------------------------------------------------------
Move Changes: Judgement:
Judgement is no longer based on what plate Arceus is holding but instead on the type Arceus is.