Some changes.
Gives back Justified to the swords of justice and hands them a Hidden Ability.

Fixes the designation of Cherry Bomb being a contact-slicing move, when it is not.

Introduces Awanezumi (1513) and Garock (1514). Moveset of Garock is currently placeholdered by a Tyrunt moveset with Flip Turn added to it. Awanezumi's moveset is partially based off of that of Marill, and is done; Garock's is not.

Places Awanezumi into NU, as it is designed for the tier.

Places Garock in OU.

Weighs Chivalmet up by 100 KG. Shelmet now properly is told to evolve into Chivalmet.

Gives Kaburion Mega Punch and Mega Kick.