Run away defiant clone
I am vehemently against this. this literally just busts so many pokemon that are otherwise bad, it dislodges SO many LC mons, as it completely wrecks the speed barrier that Voltorb and Dig/Wiglett sets. This SHOULDN'T happen. Don't create 4 types of Regieleki with actual movepools. Snubbull, a base 30 speed mon, is passing the 328 speed barrier.